Refactor crypto module and add poly1305-donna implementation
This commit refactors the crypto module by adding a new file, crypto.c, which contains functions for zeroing memory and comparing memory. It also introduces a new header file, poly1305-donna.h, which provides an implementation of the Poly1305 cryptographic hash function. These changes improve the security and performance of the crypto module.